add pixel_macro for generate code
