+++ games/snake/Cargo.toml
@@ -22,6 +22,7 @@ wasm-bindgen-futures = "0.4.42"
 web-sys = "0.3.4"
 wasm-logger = "0.2.0"
 rust_pixel = { path = "../../rust-pixel", default-features = false }
+pixel_macro = { path = "../../pixel_macro" }
 
 [features]
 sdl = ["rust_pixel/sdl"]
