+++ games/snake/src/main.rs
@@ -1,6 +1,5 @@
 use wasm_bindgen::prelude::*;
 use snake::run;
-
 fn main() -> Result<(), JsValue> {
     run()
 }
