+++ games/tetris/src/render.rs
@@ -1,5 +1,3 @@
-use std::process::Termination;
-
 use crate::model::TetrisModel;
 use tetris_lib::constant::*;
 //use std::fs::File;
