+++ pixel_macro/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "pixel_macro"
+version = "0.1.0"
+edition = "2021"
+authors.workspace = true
+description.workspace = true
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1.0"
+quote = "1.0"
+syn = { version = "1.0", features = ["full"] }
